move to cloud/bqfake
