+++ cloud/bqfake/client.go
